+++ xcube_geodb/core/geodb.py
@@ -80,6 +80,32 @@ class GeoDBError(ValueError):
 
 # noinspection PyShadowingNames,PyUnusedLocal
 class GeoDBClient(object):
+    """
+    Constructing the geoDB client. Dpending on teh setup it will automatically setup credentials from
+    environment variables. The user can also pass credentials into the constructor.
+
+    Args:
+        server_url (str): The URL of the PostGrest Rest API service
+        server_port (str): The port to the PostGrest Rest API service
+        dotenv_file (str): Name of the dotenv file [.env] to set client IDs and secrets
+        client_secret (str): Client secret (overrides environment variables)
+        client_id (str): Client ID (overrides environment variables)
+        auth_mode (str): Authentication mode [silent]. Can be 'client-credentials', 'password' and 'interactive'
+        auth_aud (str): Authentication audience
+        config_file (str): Filename that stores config info for the geodb client
+
+    Raises:
+        GeoDBError: if the auth mode does not exist
+        NotImplementedError: on auth mode interactive
+
+    Examples:
+        >>> geodb = GeoDBClient(auth_mode='client_credentials', client_id='***', client_secret='***')
+        >>> geodb.whoami
+        my_user
+    """
+
+    version = version
+
     def __init__(self,
                  server_url: Optional[str] = None,
                  server_port: Optional[int] = None,
@@ -94,19 +120,6 @@ class GeoDBClient(object):
                  config_file: str = str(Path.home()) + '/.geodb',
                  database: Optional[str] = None,
                  access_token_uri: Optional[str] = None):
-        """
-
-        Args:
-            server_url (str): The URL of the PostGrest Rest API service
-            server_port (str): The port to the PostGrest Rest API service
-            dotenv_file (str): Name of the dotenv file [.env] to set client IDs and secrets
-            client_secret (str): Client secret (overrides environment variables)
-            client_id (str): Client ID (overrides environment variables)
-            auth_mode (str): Authentication mode [silent]. Can be 'client-credentials', 'password' and 'interactive'
-            auth_aud (str): Authentication audience
-            config_file (str): Filename that stores config info for the geodb client
-        """
-
         self._dotenv_file = dotenv_file
         self._database = None
         # Access token is set here or on request
@@ -152,7 +165,7 @@ class GeoDBClient(object):
         self._config_file = config_file
 
         if self._auth_mode not in ('interactive', 'password', 'client-credentials'):
-            raise ValueError("auth_mode can only be 'interactive', 'password', or 'client-credentials'!")
+            raise GeoDBError("auth_mode can only be 'interactive', 'password', or 'client-credentials'!")
 
         if self._auth_mode == "interactive":
             raise NotImplementedError("The interactive mode has not been implemented.")
@@ -167,6 +180,10 @@ class GeoDBClient(object):
             warn(msg=msg)
 
     def _set_from_env(self):
+        """
+        Load configurations from environment variables. Overrides defaults.
+
+        """
         self._server_url = os.getenv('GEODB_API_SERVER_URL') or self._server_url
         self._server_port = os.getenv('GEODB_API_SERVER_PORT') or self._server_port
         self._auth_client_id = os.getenv('GEODB_AUTH_CLIENT_ID') or self._auth_client_id
@@ -186,10 +203,30 @@ class GeoDBClient(object):
         """
 
         Args:
-            collection (str): The name of teh collection to inspect
+            collection (str): The name of the collection to inspect
 
         Returns:
             A dictionary with collection information
+
+        Raises:
+            GeoDBError: When the collection does not exist
+
+        Examples:
+            >>> geodb.get_collection_info('my_collection')
+            {
+                'required': ['id', 'geometry'],
+                'properties': {
+                'id': {
+                    'format': 'integer', 'type': 'integer',
+                    'description': 'Note:This is a Primary Key.'
+                },
+                'created_at': {'format': 'timestamp with time zone', 'type': 'string'},
+                'modified_at': {'format': 'timestamp with time zone', 'type': 'string'},
+                'geometry': {'format': 'public.geometry(Geometry,3794)', 'type': 'string'},
+                'my_property1': {'format': 'double precision', 'type': 'number'},
+                'my_property2': {'format': 'double precision', 'type': 'number'},
+                'type': 'object'
+            }
         """
         capabilities = self.capabilities
         collection = self.database + '_' + collection
@@ -197,7 +234,7 @@ class GeoDBClient(object):
         if collection in capabilities['definitions']:
             return capabilities['definitions'][collection]
         else:
-            raise ValueError(f"Table {collection} does not exist.")
+            raise GeoDBError(f"Table {collection} does not exist.")
 
     @deprecated_func(msg='Use get_my_collections')
     def get_collections(self, database: Optional[str] = None):
@@ -206,8 +243,17 @@ class GeoDBClient(object):
     def get_my_collections(self, database: Optional[str] = None) -> Sequence:
         """
 
+        Args:
+            database (str): The database to list collections from
+
         Returns:
-            An array of collection names
+            A Dataframe of collection names
+
+        Examples:
+            >>> geodb.get_my_collections()
+            	owner	                        database	                    table_name
+            0	geodb_9bfgsdfg-453f-445b-a459	geodb_9bfgsdfg-453f-445b-a459	land_use
+
         """
         payload = {'database': database}
         r = self._post(path='/rpc/geodb_get_my_collections', payload=payload)
@@ -227,8 +273,9 @@ class GeoDBClient(object):
     @property
     def database(self) -> str:
         """
+
         Returns:
-            The current namespace
+            The current database
         """
         return self._database or self.whoami
 
@@ -246,7 +293,8 @@ class GeoDBClient(object):
         """
 
         Returns:
-            A dictionary of the PostGrest REST API service's capabilities
+            A dictionary of the geoDB PostGrest REST API service's capabilities
+
         """
         return self._capabilities or self._get(path='/').json()
 
@@ -254,6 +302,13 @@ class GeoDBClient(object):
         self._auth0_login()
 
     def _auth0_login(self):
+        """
+        Interactive login
+
+        Raises:
+            GeoDBError: When nor shell is present
+            ImportError: When ipyauth ir IPython is not installed
+        """
         try:
             from ipyauth import ParamsAuth0, Auth
             import IPython
@@ -270,7 +325,7 @@ class GeoDBClient(object):
         self._ipython_shell = IPython.get_ipython()
 
         if self._ipython_shell is None:
-            raise ValueError("You do not seem to be in an interactive ipython session. Interactive login cannot "
+            raise GeoDBError("You do not seem to be in an interactive ipython session. Interactive login cannot "
                              "be used.")
 
         auth_params = ParamsAuth0(dotenv_file=auth0_config_file, dotenv_folder=auth0_config_folder)
@@ -285,10 +340,11 @@ class GeoDBClient(object):
 
     def refresh_config_from_env(self, dotenv_file: str = ".env", use_dotenv: bool = False):
         """
-
+        Refresh the configuration from environment variables. The variables can be preset by a dotenv file.
         Args:
-            dotenv_file: A dotenv config file
-            use_dotenv: Whether to use dotenv. Might be useful if the configuration is set externally.
+            dotenv_file (str): A dotenv config file
+            use_dotenv (bool): Whether to use a dotenv file.
+
         """
         if use_dotenv:
             self._dotenv_file = find_dotenv(filename=dotenv_file)
@@ -302,11 +358,11 @@ class GeoDBClient(object):
         """
 
         Args:
-            headers: Request headers. Allows Overriding common header entries.
-            path: API path
-            payload: Post body as Dict. Will be dumped to JSON
-            params: Request parameters
-            raise_for_status: raise or not if status is not 200-299
+            headers [Optional[Dict]]: Request headers. Allows Overriding common header entries.
+            path (str): API path
+            payload (Union[Dict, Sequence]): Post body as Dict. Will be dumped to JSON
+            params Optional[Dict]: Request parameters
+            raise_for_status (bool): raise or not if status is not 200-299 [True]
         Returns:
             requests.models.Response: A Request object
 
@@ -338,9 +394,9 @@ class GeoDBClient(object):
         """
 
         Args:
-            headers: Request headers. Allows Overriding common header entries.
-            path: API path
-            params: Request parameters
+            headers (Optional[Dict]): Request headers. Allows Overriding common header entries.
+            path (str): API path
+            params (Optional[Dict]): Request parameters
 
         Returns:
             requests.models.Response: A Request object
@@ -367,9 +423,9 @@ class GeoDBClient(object):
         """
 
         Args:
-            headers: Request headers. Allows Overriding common header entries.
-            path: API path
-            params: Request parameters
+            headers (Optional[Dict]): Request headers. Allows Overriding common header entries.
+            path (str): API path
+            params (Optional[Dict]): Request parameters
 
         Returns:
             requests.models.Response: A Request object
@@ -395,10 +451,10 @@ class GeoDBClient(object):
         """
 
         Args:
-            headers: Request headers. Allows Overriding common header entries.
-            payload:
-            path: API path
-            params: Request parameters
+            headers (Optional[Dict]): Request headers. Allows Overriding common header entries.
+            payload (Union[Dict, Sequence]): Post body as Dict. Will be dumped to JSON
+            path (str): API path
+            params (Optional[Dict]): Request parameters
 
         Returns:
             requests.models.Response: A Request object
@@ -421,16 +477,26 @@ class GeoDBClient(object):
         return r
 
     def logout(self):
+        """
+        Log a user off the geoDB.
+
+        """
         self._auth_access_token = ''
         os.remove(self._config_file)
 
     def get_my_usage(self, pretty=True) -> Dict:
         """
+        Get my geoDB data usage.
+
         Args:
-            pretty: Whether to return in human readable form or in bytes
+            pretty (bool): Whether to return in human readable form or in bytes
 
         Returns:
             A dict containing the usage in bytes (int) or as a human readable string
+
+        Example:
+            >>> geodb.get_my_usage(True)
+            {'usage': '6432 kB'}
         """
         payload = {'pretty': pretty} if pretty else {}
         r = self._post(path='/rpc/geodb_get_my_usage', payload=payload)
@@ -442,6 +508,23 @@ class GeoDBClient(object):
                                         crs: int = 4326,
                                         database: Optional[str] = None,
                                         **kwargs) -> Optional[Collections]:
+        """
+        Creates a collection only if the collection does not exist already.
+
+        Args:
+            collection (str): The name of the collection to be created
+            properties (Dict): Properties to be added to teh collection
+            crs (int): projection
+            database (str): The database the collection is to be created in [current database]
+            kwargs: Placeholder for deprecated parameters
+
+        Returns:
+            Collection:  Collection info id operation succeeds
+            None: If operation fails
+
+        Examples:
+            See create_collection for an example
+        """
         exists = self.collection_exists(collection=collection, database=database)
         if not exists:
             return self.create_collection(collection=collection,
@@ -454,6 +537,20 @@ class GeoDBClient(object):
     def create_collections_if_not_exist(self,
                                         collections: Dict,
                                         database: Optional[str] = None, **kwargs) -> Collections:
+        """
+        Creates collections only if collections do not exist already.
+
+        Args:
+            collections (Dict): The name of the collection to be created
+            database (str): The database the collection is to be created in [current database]
+            kwargs: Placeholder for deprecated parameters
+
+        Returns:
+            List of Collections: List of informations about created collections
+
+        Examples:
+            See create_collections for examples
+        """
         res = dict()
         for collection in collections:
             exists = self.collection_exists(collection=collection, database=database)
@@ -470,11 +567,11 @@ class GeoDBClient(object):
                            clear: bool = False,
                            **kwargs) -> Union[Collections, Message]:
         """
-
+        Create collections from a dictionary
         Args:
-            clear: Delete collections prioer to creation
-            collections: A dictionalry of collections
-            database:
+            clear (bool): Delete collections prioer to creation
+            collections (Dict): A dictionalry of collections
+            database (str): Database to use for creating the collection
 
         Returns:
             bool: Success
@@ -517,13 +614,14 @@ class GeoDBClient(object):
                           clear: bool = False,
                           **kwargs) -> Collections:
         """
+        Create collections from a dictionary
 
         Args:
-            clear: Whether to delete existing collections
-            properties: Property definitions for the collection
-            collection: Collection to be created
+            collection (str): Name of the collection to be created
+            clear (bool): Whether to delete existing collections
+            properties (Dict): Property definitions for the collection
+            database (str): Database to use for creating the collection
             crs: sfdv
-            database: name of database if not default [user name].
 
         Returns:
             bool: Success
@@ -550,8 +648,9 @@ class GeoDBClient(object):
         """
 
         Args:
-            collection: Collection to be dropped
-            database:
+            collection (str): Name of the collection to be dropped
+            database (str): The database teh colections resides in [current database]
+            kwargs: Placeholder for deprecated options
 
         Returns:
             bool: Success
@@ -569,8 +668,9 @@ class GeoDBClient(object):
         """
 
         Args:
-            database:
-            collections: Collections to be dropped
+            database (str): The database teh colections resides in [current database]
+            collections (Sequence[str]): Collections to be dropped
+            kwargs: Placeholder for deprecated options
 
         Returns:
             Message
@@ -597,12 +697,20 @@ class GeoDBClient(object):
         """
 
         Args:
-            collection: Collection to grant access to
-            usr: User to grant access to
-            database: The namespace to grant access to [public]. By default, public access is granted
+            collection (str): Collection name to grant access to
+            usr (str): Username to grant access to
+            database (str): The database the collection resides in
+            kwargs: Placeholder for deprecated options
 
         Returns:
             bool: Success
+
+        Raises:
+            HttpError: when http request fails
+
+        Examples:
+            >>> geodb.grant_access_to_collection('[Collection]', '[User who gets access]')
+            Access granted on Collection to User who gets access}
         """
         database = database or self.database
         dn = f"{database}_{collection}"
@@ -612,6 +720,17 @@ class GeoDBClient(object):
         return Message(f"Access granted on {collection} to {usr}")
 
     def rename_collection(self, collection: str, new_name: str, database: Optional[str] = None):
+        """
+
+        Args:
+            collection (str): The name of the collection to be renamed
+            new_name (str):The new name of the collection
+            database (str): The database the collection resides in
+
+        Raises:
+            HttpError: When request fails
+        """
+
         database = database or self._database
 
         old_dn = f"{database}_{collection}"
@@ -620,6 +739,18 @@ class GeoDBClient(object):
         self._post(path='/rpc/geodb_rename_collection', payload={'collection': old_dn, 'new_name': new_dn})
 
     def move_collection(self, collection: str, new_database: str, database: Optional[str] = None):
+        """
+        Move a collection from one database to another
+
+        Args:
+            collection (str): The name of the collection to be renamed
+            new_database (str): The database the collection will be moved to
+            database (str): The database the collection resides in
+
+        Examples:
+            >>> geodb.move_collection('[Collection]', '[New Database]')
+        """
+
         database = database or self._database
         old_dn = f"{database}_{collection}"
         new_dn = f"{new_database}_{collection}"
@@ -627,6 +758,18 @@ class GeoDBClient(object):
         self._post(path='/rpc/geodb_rename_collection', payload={'collection': old_dn, 'new_name': new_dn})
 
     def copy_collection(self, collection: str, new_collection: str, new_database: str, database: Optional[str] = None):
+        """
+
+        Args:
+            collection (str): The name of the collection to be copied
+            new_collection (str): The new name of the collection
+            database (str): The database the collection resides in [current database]
+            new_database (str): The database the collection will be copied to
+
+        Examples:
+            >>> geodb.copy_collection('[Collection]', '[New Collection]')
+        """
+
         database = database or self._database
         from_dn = f"{database}_{collection}"
         to_dn = f"{new_database}_{new_collection}"
@@ -635,13 +778,16 @@ class GeoDBClient(object):
 
     def publish_collection(self, collection: str, database: Optional[str] = None) -> Message:
         """
-
+        Publish a collection. The collection will bew accessible by all users in the geoDB.
         Args:
-            database:
-            collection: Collection to grant access to
+            database (str): The database the collection resides in [current database]
+            collection (str): The name of the collection that will be made public
 
         Returns:
-            str: Message
+            Message: Message whether operation succeeded
+
+        Examples:
+            >>> geodb.publish_collection('[Collection]')
         """
         try:
             database = database or self.database
@@ -655,12 +801,16 @@ class GeoDBClient(object):
 
     def unpublish_collection(self, collection: str, database: Optional[str] = None) -> Message:
         """
-
+        Revoke public access to a collection. The collection will nor be accessible by all users in the geoDB.
         Args:
-            collection: Collection to grant access to
+            database (str): The database the collection resides in [current database]
+            collection (str): The name of the collection that will be removed from public access
 
         Returns:
-            str: Message
+            Message: Message whether operation succeeded
+
+        Examples:
+            >>> geodb.unpublish_collection('[Collection]')
         """
 
         try:
@@ -678,14 +828,14 @@ class GeoDBClient(object):
     def revoke_access_from_collection(self, collection: str, usr: str, database: Optional[str] = None,
                                       **kwargs) -> Message:
         """
-
+        Revoke access from a collection
         Args:
-            collection: Collection to grant access to
-            usr: User to revoke access from
-            database: The user to revoke access from [public].
+            collection (str): Name of the collection
+            usr (str): User to revoke access from
+            database (str): The database the collection resides in [current database]
 
         Returns:
-            bool: Success
+            Message: Whether operation has succeeded
         """
         database = database or self.database
         dn = f"{database}_{collection}"
@@ -702,10 +852,13 @@ class GeoDBClient(object):
 
     def list_my_grants(self) -> DataFrame:
         """
+        List the access grants the current user has granted
 
         Returns:
-            A list of the current user's collection grants
+            DataFrame: A list of the current user's access grants
 
+        Raises:
+            GeoDBError: If access to geoDB fails
         """
         r = self._post(path='/rpc/geodb_list_grants', payload={})
         try:
@@ -721,14 +874,15 @@ class GeoDBClient(object):
     def add_property(self, collection: str, prop: str, typ: str, database: Optional[str] = None, **kwargs) -> Message:
         """
         Add a property to an existing collection
+
         Args:
-            collection: Collection to add a property to
-            prop: Property name
-            typ: Type of property (Postgres type)
-            database:
+            collection (str): The name of the collection to add a property to
+            prop (str): Property name
+            typ (str): The data type of the property (Postgres type)
+            database (str): The database the collection resides in [current database]
 
         Returns:
-            Success Message
+            Message: Success Message
 
         Examples:
             >>> geodb = GeoDBClient()
@@ -740,13 +894,14 @@ class GeoDBClient(object):
     @deprecated_kwarg('namespace', 'database')
     def add_properties(self, collection: str, properties: Dict, database: Optional[str] = None, **kwargs) -> Message:
         """
+        Add properties to a collection
 
         Args:
-            collection: Collection to add properties to
-            properties: Property definitions as json array
-            database:
+            collection (str): The name of the collection to add properties to
+            properties (Dict): Property definitions as dictionary
+            database (str): The database the collection resides in [current database]
         Returns:
-            bool: Success
+            Message: Whether the operation succeeded
 
         Examples:
             >>> properties = {'[MyName1]': '[PostgresType1]', '[MyName2]': '[PostgresType2]'}
@@ -766,14 +921,14 @@ class GeoDBClient(object):
     @deprecated_kwarg('namespace', 'database')
     def drop_property(self, collection: str, prop: str, database: Optional[str] = None, **kwargs) -> Message:
         """
-
+        Drop a property from a collection
         Args:
-            collection: Collection to drop the property from
-            prop: Property to delete
-            database:
+            collection (str): The name of teh collection to drop the property from
+            prop (str): The property to delete
+            database (str): The database the collection resides in [current database]
 
         Returns:
-            bool: Success
+            Message: Whether the operation succeeded
 
         Examples:
             >>> geodb = GeoDBClient()
@@ -786,13 +941,13 @@ class GeoDBClient(object):
     def drop_properties(self, collection: str, properties: Sequence[str], database: Optional[str] = None,
                         **kwargs) -> Message:
         """
-
+        Drop poperties from a collection
         Args:
-            collection: Collection to delete properties from
-            properties: A json object containing the property definitions
-            database:
+            collection (str): The name of the collection to delete properties from
+            properties (Dict): A dictionary containing the property definitions
+            database (str): The database the collection resides in [current database]
         Returns:
-            bool: Success
+            Message: Whether the operation succeeded
 
         Examples:
             >>> geodb = GeoDBClient()
@@ -819,10 +974,11 @@ class GeoDBClient(object):
     @deprecated_kwarg('namespace', 'database')
     def get_properties(self, collection: str, database: Optional[str] = None, **kwargs) -> DataFrame:
         """
+        Get a list of properties of a collection
 
         Args:
-            collection: Collection to retrieve a list of properties from
-            database:
+            collection (str): The name of the collection to retrieve a list of properties from
+            database (str): The database the collection resides in [current database]
 
         Returns:
             DataFrame: A list of properties
@@ -842,9 +998,13 @@ class GeoDBClient(object):
 
     def create_database(self, database: str) -> Message:
         """
+        Create a database
+
+        Args:
+            database (str): The name of teh database to be created
 
         Returns:
-            DataFrame: A list of collections the user owns
+            Message: A message about the success or failure of the operation
 
         """
 
@@ -856,9 +1016,13 @@ class GeoDBClient(object):
 
     def truncate_database(self, database: str) -> Message:
         """
+        Delete all tables in the given database
+
+        Args:
+            database (str): The name of teh database to be created
 
         Returns:
-            DataFrame: A list of collections the user owns
+            Message: A message about the success or failure of the operation
 
         """
 
@@ -870,6 +1034,7 @@ class GeoDBClient(object):
 
     def get_my_databases(self):
         """
+        Get a list of databases the current user owns
 
         Returns:
             DataFrame: A list of databases the user owns
@@ -880,10 +1045,16 @@ class GeoDBClient(object):
 
     def database_exists(self, database: str) -> bool:
         """
-        Checkes whether a database exists
+        Checks whether a database exists
+
+        Args:
+            database (str): The name of the database to be checked
 
         Returns:
             bool: database exists
+
+        Raises:
+            HttpError: If request fails
 
         """
 
@@ -893,13 +1064,15 @@ class GeoDBClient(object):
     @deprecated_kwarg('namespace', 'database')
     def delete_from_collection(self, collection: str, query: str, database: Optional[str] = None, **kwargs) -> Message:
         """
-
+        Delete
         Args:
-            collection: Collection to delete from  
-            query: Filter which records to delete. Follow the http://postgrest.org/en/v6.0/api.html query convention.
-            database:
+            collection (str): The name of the collection to delete rows from
+            database (str): The name of the database to be checked
+            query (str): Filter which records to delete. Follow the http://postgrest.org/en/v6.0/api.html query
+            convention.
+            kwargs: PLaceholder for deprecated options
         Returns:
-            bool: Success
+            Message: Whether the operation has succeeded
 
         Examples:
             >>> geodb = GeoDBClient()
@@ -919,14 +1092,21 @@ class GeoDBClient(object):
     def update_collection(self, collection: str, values: Dict, query: str, database: Optional[str] = None,
                           **kwargs) -> Message:
         """
+        Update data in a collection by a query
 
         Args:
-            collection: Collection to be updated
-            values: Values to update
-            query: Filter which values to be updated. Follow the http://postgrest.org/en/v6.0/api.html query convention.
-            database:
+            collection (str): The name of the collection to be updated
+            database (str): The name of the database to be checked
+            values (Dict): Values to update
+            query (str): Filter which values to be updated. Follow the http://postgrest.org/en/v6.0/api.html query
+            convention.
         Returns:
             Message: Success
+
+        Raises:
+            GeoDBError: if the values is not a Dict or request fails
+        Example:
+
         """
 
         database = database or self.database
@@ -938,7 +1118,7 @@ class GeoDBClient(object):
             if 'id' in values.keys():
                 del values['id']
         else:
-            raise ValueError(f'Format {type(values)} not supported.')
+            raise GeoDBError(f'Format {type(values)} not supported.')
 
         try:
             self._patch(f'/{dn}?{query}', payload=values)
@@ -989,21 +1169,26 @@ class GeoDBClient(object):
                                **kwargs) \
             -> Message:
         """
+        Insert data into a collection
 
         Args:
-            database:
-            collection: Collection to be inserted to
-            values: Values to be inserted
-            upsert: Whether the insert shall replace existing rows (by PK)
-            crs: crs (in the form of an SRID) of the geometries. If not present, tssi method will attempt to guess it
-            from the GeoDataFrame input. Must be in sync with the target collection in the GeoDatabase
-            max_transfer_chunk_size: Maximum number of rows per chunk to be sent to the geodb.
+            collection (str): Collection to be inserted to
+            database (str): The name of the database teh collection resides in [current database]
+            values (GeoDataFrame): Values to be inserted
+            upsert (bool): Whether the insert shall replace existing rows (by PK)
+            crs (int): crs (in the form of an SRID) of the geometries. If not present, the method will attempt
+            guessing it from the GeoDataFrame input. Must be in sync with the target collection in the GeoDatabase
+            max_transfer_chunk_size (int): Maximum number of rows per chunk to be sent to the geodb.
 
         Raises:
             ValueError: When crs is not given and cannot be guessed from the GeoDataFrame
+            GeoDBError: If the values are not in format Dict
 
         Returns:
             bool: Success
+
+        Example:
+
         """
 
         # self._collection_exists(collection=collection)
@@ -1067,18 +1252,18 @@ class GeoDBClient(object):
                                database: Optional[str] = None,
                                **kwargs) -> GeoDataFrame:
         """
+        Query the database by a boundnng box.
 
         Args:
-            collection: Table to get
+            collection (str): The name of teh collection to be quried
             bbox (int, int, int, int): minx, maxx, miny, maxy
-            comparison_mode: Filter mode. Can be 'contains' or 'within' ['contains']
-            bbox_crs: Projection code. [4326]
-            op: Operator for where (AND, OR) ['AND']
-            where: Additional SQL where statement
-            limit: Limit for paging
-            offset: Offset (start) of rows to return. Used in combination with lmt.
-            database: By default the API gets in the user's own namespace. To access
-                       collections the user has grant set the namespace accordingly.
+            comparison_mode (str): Filter mode. Can be 'contains' or 'within' ['contains']
+            bbox_crs (int): Projection code. [4326]
+            op (str): Operator for where (AND, OR) ['AND']
+            where (str): Additional SQL where statement to further filter the collection
+            limit (int): The maximum number of rows to be returned
+            offset (int): Offset (start) of rows to return. Used in combination with limit.
+            database (str): The name of the database teh collection resides in [current database]
 
         Returns:
             A GeoPandas Dataframe
@@ -1125,12 +1310,12 @@ class GeoDBClient(object):
     def head_collection(self, collection: str, num_lines: int = 10, database: Optional[str] = None, **kwargs) -> \
             Union[GeoDataFrame, DataFrame]:
         """
+        Get the first num_lines of a collection
 
         Args:
-            collection: The collection's name
-            num_lines: The number of line to return
-            database: By default the API gets in the user's own database. To access
-                       collections the user has grant set the namespace accordingly.
+            collection (str): The collection's name
+            num_lines (int): The number of line to return
+            database (str): The name of the database teh collection resides in [current database]
 
         Returns:
             GeoDataFrame or DataFrame: results
@@ -1151,12 +1336,12 @@ class GeoDBClient(object):
                        **kwargs) \
             -> Union[GeoDataFrame, DataFrame]:
         """
+        Query a collection
 
         Args:
-            collection: The collection's name
-            query: A query. Follow the http://postgrest.org/en/v6.0/api.html query convention.
-            database: By default the API gets in the user's own namespace. To access
-                       collections the user has grant set the namespace accordingly.
+            collection (str): The collection's name
+            query (str): A query. Follow the http://postgrest.org/en/v6.0/api.html query convention.
+            database (str): The name of the database teh collection resides in [current database]
 
         Returns:
             GeoDataFrame or DataFrame: results
@@ -1214,15 +1399,14 @@ class GeoDBClient(object):
         """
 
         Args:
-            collection(str): Collection to query
-            select(str): Properties (columns) to return. Can contain aggregation functions
-            where(Optional[str]): SQL WHERE statement
-            group(Optional[str]): SQL GROUP statement
-            order(Optional[str]): SQL ORDER statement
-            limit(Optional[int]): Limit for paging
-            offset(Optional[int]): Offset (start) of rows to return. Used in combination with limit.
-            database: By default the API gets in the user's own namespace. To access
-                       collections the user has grant set the namespace accordingly.
+            collection (str): The name of the collection to query
+            select (str): Properties (columns) to return. Can contain aggregation functions
+            where (Optional[str]): SQL WHERE statement
+            group (Optional[str]): SQL GROUP statement
+            order (Optional[str]): SQL ORDER statement
+            limit (Optional[int]): Limit for paging
+            offset (Optional[int]): Offset (start) of rows to return. Used in combination with limit.
+            database (str): The name of the database teh collection resides in [current database]
 
         Returns:
             GeoDataFrame or DataFrame: results
@@ -1266,13 +1450,24 @@ class GeoDBClient(object):
     @property
     def server_url(self) -> str:
         """
+        Get URL of the geoDb server
 
         Returns:
             str: The URL of the GeoDB REST service
         """
         return self._server_url
 
-    def get_collection_srid(self, collection: str, database: Optional[str] = None):
+    def get_collection_srid(self, collection: str, database: Optional[str] = None) -> Optional[str]:
+        """
+        Get the SRID of a collection
+
+        Args:
+            collection (str): The collection's name
+            database (str): The name of the database the collection resides in [current database]
+
+        Returns:
+            The name of the SRID
+        """
         tab_prefix = database or self.database
         dn = f"{tab_prefix}_{collection}"
 
@@ -1285,17 +1480,15 @@ class GeoDBClient(object):
 
         return None
 
-    def _df_from_json(self, js: json, srid: Optional[int] = None) -> Union[GeoDataFrame, DataFrame]:
+    def _df_from_json(self, js: Dict, srid: Optional[int] = None) -> Union[GeoDataFrame, DataFrame]:
         """
         Converts wkb geometry string to wkt from a PostGrest json result
         Args:
-            js: Json string. Will convert geometry to
+            js (Dict): The geometries to be converted
+            srid (Optional[int]):.
 
         Returns:
-            GeoDataFrame
-
-        Raises:
-            ValueError: When the geometry field is missing
+            GeoDataFrame, DataFrame
 
         """
         if js is None:
@@ -1316,7 +1509,7 @@ class GeoDBClient(object):
         """
 
         Args:
-            path: PostGrest API path
+            path (str): PostgREST API path
 
         Returns:
             str: Full URL and path
@@ -1331,10 +1524,10 @@ class GeoDBClient(object):
         """
 
         Args:
-            d: A row of thePostGrest result
+            d: A row of the PostgREST result
 
         Returns:
-            Dict: A row of thePostGrest result with its geometry converted from wkb to wkt
+            Dict: A row of the PostgREST result with its geometry converted from wkb to wkt
         """
 
         if 'geometry' in d:
@@ -1344,16 +1537,20 @@ class GeoDBClient(object):
     @property
     def auth_access_token(self) -> str:
         """
+        Get the user's access token from
 
         Returns:
             The current authentication access_token
+
+        Raises:
+            GeoDBError on missing ipython shell
         """
 
         # Get token from cache
         if self._auth_access_token is not None:
             token = self._auth_access_token
         else:
-            token = self._get_token_from_file()
+            token = self._get_token_from_cache()
 
         if token:
             return token
@@ -1369,11 +1566,22 @@ class GeoDBClient(object):
         return token
 
     def refresh_auth_access_token(self):
+        """
+        Refresh the authentication token
+
+        """
         self._auth_access_token = None
 
-    def _get_token_from_file(self) -> Union[str, bool]:
+    def _get_token_from_cache(self) -> Union[str, bool]:
+        """
+        Load a token from a cache file
+
+        Returns:
+            An access token or false on failure
+        """
         if os.path.isfile(self._config_file):
             with open(self._config_file, 'r') as f:
+                # noinspection PyBroadException
                 try:
                     cfg_data = json.load(f)
 
@@ -1393,25 +1601,55 @@ class GeoDBClient(object):
                     if 'access_token' in cfg_data['data']:
                         return cfg_data['data']['access_token']
                 except Exception as e:
-                    print(str(e))
                     return False
 
         return False
 
-    def _raise_for_invalid_password_cfg(self):
+    def _raise_for_invalid_password_cfg(self) -> bool:
+        """
+        Raise when the password configuration is wrong
+
+        Returns:
+             True on success
+
+        Raises:
+            GeoDBError on invalid configuration
+        """
         if self._auth_username and self._auth_password and self._auth_client_id and self._auth_client_secret \
-                and self._auth_aud:
+                and self._auth_aud == "password":
             return True
         else:
             raise GeoDBError("System: Invalid password flow configuration")
 
-    def _raise_for_invalid_client_credentials_cfg(self):
-        if self._auth_client_id and self._auth_client_secret and self._auth_aud:
+    def _raise_for_invalid_client_credentials_cfg(self) -> bool:
+        """
+        Raise when the client-credentials configuration is wrong
+
+        Returns:
+             True on success
+
+        Raises:
+            GeoDBError on invalid configuration
+        """
+        if self._auth_client_id and self._auth_client_secret and self._auth_aud == "client-credentials":
             return True
         else:
-            raise GeoDBError("System: Invalid password client_credentials configuration")
+            raise GeoDBError("System: Invalid client_credentials configuration.")
 
-    def _get_geodb_client_credentials_access_token(self, token_uri: str = "/oauth/token", is_json: bool = True):
+    def _get_geodb_client_credentials_access_token(self, token_uri: str = "/oauth/token", is_json: bool = True) -> str:
+        """
+        Get access token from client credentials
+
+        Args:
+            token_uri (str): oauth2 token URI
+            is_json: whether the request has to be of content type json
+
+        Returns:
+             An access token
+
+        Raises:
+            GeoDBError, HttpError
+        """
         payload = {}
 
         if self._auth_mode == "client-credentials":
@@ -1452,24 +1690,35 @@ class GeoDBClient(object):
         try:
             return data['access_token']
         except KeyError:
-            raise ValueError("The authorization request did net return an access token. Please contact helpdesk.")
+            raise GeoDBError("The authorization request did net return an access token.")
 
     # noinspection PyMethodMayBeStatic
     def _validate(self, df: gpd.GeoDataFrame) -> bool:
         """
+        Validate whether a GeoDataFrame is a valid geoDB frame
 
         Args:
-            df: A geopands dataframe to validate columns. Must be "raba_pid", 'raba_id', 'd_od' or 'geometry'
+            df (GeoDataFrame): Frame to check
 
         Returns:
-            bool whether validation succeeds
+            Whether id and geometry are properties in the frame
         """
+
         cols = set([x.lower() for x in df.columns])
         valid_columns = {'id', 'geometry'}
 
         return len(list(valid_columns - cols)) == 0
 
-    def collection_exists(self, collection: str, database: str):
+    def collection_exists(self, collection: str, database: str) -> bool:
+        """
+        Checks whether a collection exists
+
+        Args:
+            collection (str): The collection's name
+            database (str): The name of the database the collection resides in [current database]
+        Returns:
+             Whether the collection exists
+        """
         tab_prefix = database or self.database
         dn = f"{tab_prefix}_{collection}"
         if collection in self.capabilities['definitions']:
@@ -1480,34 +1729,35 @@ class GeoDBClient(object):
         """
 
         Args:
-            collection: A table name to check
+            collection (str): Name of the collection
 
         Returns:
-            bool whether the table exists
+            Whether the collection exists
 
+        Raises:
+            GeoDBError if the collection does not exist
         """
         if collection in self.capabilities['definitions']:
             return True
         else:
-            raise ValueError(f"Collection {collection} does not exist")
+            raise GeoDBError(f"Collection {collection} does not exist")
 
     def _raise_for_stored_procedure_exists(self, stored_procedure: str) -> bool:
         """
 
         Args:
-            stored_procedure: Name of stored pg procedure
+            stored_procedure (str): Name of the stored procedure
 
         Returns:
-            bool whether the stored procedure exists in DB
+            Whether the stored procedure exists
+
+        Raises:
+            GeoDBError if the stored procedure does not exist
         """
         if f"/rpc/{stored_procedure}" in self.capabilities['paths']:
             return True
         else:
-            raise ValueError(f"Stored procedure {stored_procedure} does not exist")
-
-    @staticmethod
-    def version():
-        return version
+            raise GeoDBError(f"Stored procedure {stored_procedure} does not exist")
 
     @staticmethod
     def setup(host: Optional[str] = None,
@@ -1517,7 +1767,7 @@ class GeoDBClient(object):
               dbname: Optional[str] = None,
               conn: Optional[any] = None):
         """
-            Sets up  the datase. Needs DB credentials and the database user requires CREATE TABLE/FUNCTION grants.
+            Sets up  the database. Needs DB credentials and the database user requires CREATE TABLE/FUNCTION grants.
         """
         host = host or os.getenv('GEODB_DB_HOST')
         port = port or os.getenv('GEODB_DB_PORT')
@@ -1540,6 +1790,13 @@ class GeoDBClient(object):
         conn.commit()
 
     def list_users(self):
+        """
+        List the users in a database. Needs DB credentials and the database user requires CREATE TABLE/FUNCTION grants.
+
+        Returns:
+            A list of users
+
+        """
         r = self._get(path='/rpc/geodb_list_users')
         if r.status_code == 200:
             js = r.json()[0]['src'][0]
@@ -1549,7 +1806,17 @@ class GeoDBClient(object):
 
         return None
 
-    def register_user(self, user_name: str, password: str):
+    def register_user(self, user_name: str, password: str) -> bool:
+        """
+        Register a user to the geoDB. Needs DB credentials and the database user requires CREATE TABLE/FUNCTION grants.
+
+        Args:
+            user_name (str): The user name
+            password (str): The password of the user
+        Returns:
+            Whether registering the user was successful
+
+        """
         payload = {
             'user_name': user_name,
             'password': password